Build fix / fix for ticket #99.

+++ mapiproxy/modules/mpm_cache.h
@@ -107,7 +107,7 @@ NTSTATUS	mpm_cache_ldb_add_stream(struct mpm_cache *, struct ldb_context *, stru
 NTSTATUS	mpm_cache_stream_open(struct mpm_cache *, struct mpm_stream *);
 NTSTATUS	mpm_cache_stream_close(struct mpm_stream *);
 NTSTATUS	mpm_cache_stream_write(struct mpm_stream *, uint16_t, uint8_t *);
-NTSTATUS	mpm_cache_stream_read(struct mpm_stream *, uint32_t, size_t *, uint8_t **);
+NTSTATUS	mpm_cache_stream_read(struct mpm_stream *, size_t, size_t *, uint8_t **);
 NTSTATUS	mpm_cache_stream_reset(struct mpm_stream *);
 
 __END_DECLS
